Support empty documents
Whenever `transcend.io` is in a TCM bundle's allowed domains list, we end up injecting airgap.js into the TCM sync frame. The TCM sync frame contains an 'empty' document post-load, and this UI is trying to call `null.append()` in that case.

We also made a change upstream for airgap.js to prevent this from happening for our sync endpoints.